// src/types.d.ts
export type StateName = string;
export type EventName = string;
export type EventPayload = any;
export type OnEnter = (meta: StateChangeHookMeta, misc: {handlePrivate: HandlePrivate}) => any;
export type OnExit = (meta: StateChangeHookMeta) => any;
export type StateChangeHookMeta = {
  fromStateName: StateName,
  toStateName: StateName,
  eventPayload: EventPayload,
  changeStatePayload: ChangeStatePayload,
};
export type ChangeStateClosure = (toStateName: StateName, changeStatePayload?: ChangeStatePayload) => void;
export type ChangeStatePayload = any;
export type Handler = (
  changeState: ChangeStateClosure,
  eventPayloadObj: { eventPayload: EventPayload },
  misc: {
    handlePrivate: HandlePrivate
  }
) => any;
export type HandlerObj = {
  fn: Handler;
  isPrivate: boolean;
};
export type HandlePrivate = (eventName: EventName, eventPayload: EventPayload) => Promise<HandleMeta>;
export type HandlerMiddleware = (meta: any) => any;
// ...
export type HandleStateChangeResult = {
  fromStateName: StateName;
  toStateName: StateName;
  enter: ReturnType<State['enter']>;
  exit: ReturnType<State['exit']>;
};

export type HandleMeta = {
  hasHandler: boolean;
  beforeHandleResult: ReturnType<HandlerMiddleware>;
  afterHandleResult: ReturnType<HandlerMiddleware>;
  handlerResult: ReturnType<Handler>;
  changeStateResult: HandleStateChangeResult;
  isPrivate: boolean;
};
// ...
export type StateConfig = {
  /** The name of the state. */
  stateName: StateName,
  /** Callback to be called on entering state. */
  onEnter?: OnEnter,
  /** Callback to be called on exiting state. */
  onExit?: OnExit,
  /** Callbacks keyed by state to be called when entering this state, when arriving from the keyed state. */
  onEnterFrom?: {[KStateName in StateName]: OnEnter},
  /** Callbacks keyed by state to be called when exiting this state, when leaving to the keyed state. */
  onExitTo?: {[KStateName in StateName]: OnExit},
  /** Event handlers which may be triggered from the state machine. */
  handlers?: {[KEventName in EventName]: Handler},
  /** Event handlers which may only be triggered from the state's onEnter function. */
  privateHandlers?: {[KEventName in EventName]: Handler},
  /** Callback to be called prior to handlers. */
  beforeHandle?: HandlerMiddleware,
  /** Callback to be called after handlers. */
  afterHandle?: HandlerMiddleware,
};

/**
 * The namespace for the State\ class.
 */
export type StatePrivateNamespace = {
  stateName: StateName;
  onEnter: OnEnter;
  onExit: OnExit;
  onEnterFrom: {[KStateName in StateName]: OnEnter};
  onExitTo: {[KStateName in StateName]: OnExit};
  handlers: {[KEventName in EventName]: HandlerObj};
  handlerMiddleware: {
    before?: HandlerMiddleware;
    after?: HandlerMiddleware;
  };
};